logins: split the DB from the sync engine
